tests: Split out gpu-av and printf into their own classes
